Add test case to ensure classpath is reuse
